requirements.txt
numpy
Pillow
tensorflow
tqdm